+++ nova/src/relaxed_r1cs/witness.rs
@@ -1,17 +1,17 @@
 use crate::{PedersenCommitment, R1csShape};
 use zkstd::circuit::prelude::CircuitDriver;
-use zkstd::common::IntGroup;
+use zkstd::common::{Decode, Encode, IntGroup};
 use zkstd::matrix::DenseVectors;
 
 /// A type that holds a witness for a given R1CS instance
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq, Eq, Encode, Decode)]
 pub struct R1csWitness<C: CircuitDriver> {
     pub w: DenseVectors<C::Scalar>,
 }
 
 impl<C: CircuitDriver> R1csWitness<C> {
     pub fn new(shape: &R1csShape<C>, w: Vec<C::Scalar>) -> Self {
-        assert_eq!(shape.m_l_1(), w.len());
+        assert_eq!(shape.m_l_1() as usize, w.len());
         Self {
             w: DenseVectors::new(w),
         }
@@ -26,7 +26,7 @@ impl<C: CircuitDriver> R1csWitness<C> {
     }
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq, Eq, Encode, Decode)]
 pub struct RelaxedR1csWitness<C: CircuitDriver> {
     /// witness
     pub(crate) w: DenseVectors<C::Scalar>,
@@ -38,7 +38,7 @@ impl<C: CircuitDriver> RelaxedR1csWitness<C> {
     pub fn from_r1cs_witness(shape: &R1csShape<C>, witness: &R1csWitness<C>) -> Self {
         Self {
             w: witness.w.clone(),
-            e: DenseVectors::new(vec![C::Scalar::zero(); shape.m()]),
+            e: DenseVectors::new(vec![C::Scalar::zero(); shape.m() as usize]),
         }
     }
